+++ internal/xsql/conn.go
@@ -440,7 +440,10 @@ func (c *conn) Version(context.Context) (_ string, err error) {
 
 func (c *conn) IsTableExists(ctx context.Context, tableName string) (tableExists bool, err error) {
 	tableName = c.normalizePath(tableName)
-	tableExists, err = helpers.IsTableExists(ctx, c.connector.parent.Scheme(), tableName)
+	tableExists, err = helpers.IsEntryExists(ctx,
+		c.connector.parent.Scheme(), tableName,
+		scheme.EntryTable, scheme.EntryColumnTable,
+	)
 	if err != nil {
 		return false, xerrors.WithStackTrace(err)
 	}
@@ -449,7 +452,10 @@ func (c *conn) IsTableExists(ctx context.Context, tableName string) (tableExists
 
 func (c *conn) IsColumnExists(ctx context.Context, tableName, columnName string) (columnExists bool, err error) {
 	tableName = c.normalizePath(tableName)
-	tableExists, err := helpers.IsTableExists(ctx, c.connector.parent.Scheme(), tableName)
+	tableExists, err := helpers.IsEntryExists(ctx,
+		c.connector.parent.Scheme(), tableName,
+		scheme.EntryTable, scheme.EntryColumnTable,
+	)
 	if err != nil {
 		return false, xerrors.WithStackTrace(err)
 	}
@@ -478,7 +484,10 @@ func (c *conn) IsColumnExists(ctx context.Context, tableName, columnName string)
 
 func (c *conn) GetColumns(ctx context.Context, tableName string) (columns []string, err error) {
 	tableName = c.normalizePath(tableName)
-	tableExists, err := helpers.IsTableExists(ctx, c.connector.parent.Scheme(), tableName)
+	tableExists, err := helpers.IsEntryExists(ctx,
+		c.connector.parent.Scheme(), tableName,
+		scheme.EntryTable, scheme.EntryColumnTable,
+	)
 	if err != nil {
 		return nil, xerrors.WithStackTrace(err)
 	}
@@ -504,7 +513,10 @@ func (c *conn) GetColumns(ctx context.Context, tableName string) (columns []stri
 
 func (c *conn) GetColumnType(ctx context.Context, tableName, columnName string) (dataType string, err error) {
 	tableName = c.normalizePath(tableName)
-	tableExists, err := helpers.IsTableExists(ctx, c.connector.parent.Scheme(), tableName)
+	tableExists, err := helpers.IsEntryExists(ctx,
+		c.connector.parent.Scheme(), tableName,
+		scheme.EntryTable, scheme.EntryColumnTable,
+	)
 	if err != nil {
 		return "", xerrors.WithStackTrace(err)
 	}
@@ -541,7 +553,10 @@ func (c *conn) GetColumnType(ctx context.Context, tableName, columnName string) 
 
 func (c *conn) GetPrimaryKeys(ctx context.Context, tableName string) (pkCols []string, err error) {
 	tableName = c.normalizePath(tableName)
-	tableExists, err := helpers.IsTableExists(ctx, c.connector.parent.Scheme(), tableName)
+	tableExists, err := helpers.IsEntryExists(ctx,
+		c.connector.parent.Scheme(), tableName,
+		scheme.EntryTable, scheme.EntryColumnTable,
+	)
 	if err != nil {
 		return nil, xerrors.WithStackTrace(err)
 	}
@@ -565,7 +580,10 @@ func (c *conn) GetPrimaryKeys(ctx context.Context, tableName string) (pkCols []s
 
 func (c *conn) IsPrimaryKey(ctx context.Context, tableName, columnName string) (ok bool, err error) {
 	tableName = c.normalizePath(tableName)
-	tableExists, err := helpers.IsTableExists(ctx, c.connector.parent.Scheme(), tableName)
+	tableExists, err := helpers.IsEntryExists(ctx,
+		c.connector.parent.Scheme(), tableName,
+		scheme.EntryTable, scheme.EntryColumnTable,
+	)
 	if err != nil {
 		return false, xerrors.WithStackTrace(err)
 	}
@@ -701,7 +719,10 @@ func (c *conn) GetAllTables(ctx context.Context, folder string) (tables []string
 
 func (c *conn) GetIndexes(ctx context.Context, tableName string) (indexes []string, err error) {
 	tableName = c.normalizePath(tableName)
-	tableExists, err := helpers.IsTableExists(ctx, c.connector.parent.Scheme(), tableName)
+	tableExists, err := helpers.IsEntryExists(ctx,
+		c.connector.parent.Scheme(), tableName,
+		scheme.EntryTable, scheme.EntryColumnTable,
+	)
 	if err != nil {
 		return nil, xerrors.WithStackTrace(err)
 	}
@@ -728,7 +749,10 @@ func (c *conn) GetIndexes(ctx context.Context, tableName string) (indexes []stri
 
 func (c *conn) GetIndexColumns(ctx context.Context, tableName, indexName string) (columns []string, err error) {
 	tableName = c.normalizePath(tableName)
-	tableExists, err := helpers.IsTableExists(ctx, c.connector.parent.Scheme(), tableName)
+	tableExists, err := helpers.IsEntryExists(ctx,
+		c.connector.parent.Scheme(), tableName,
+		scheme.EntryTable, scheme.EntryColumnTable,
+	)
 	if err != nil {
 		return nil, xerrors.WithStackTrace(err)
 	}